+++ lib/redis/redis_preflight.sh
@@ -6,5 +6,12 @@ LOGFILE=$PWD/tmp/progress.txt
 
 # This script runs before the container is composed.
 
-# store the password in tmp/postgres/db-secret
-mv tmp/secrets/redis-secret-key tmp/redis/secret-key
+# if there isn't already a value, store the password in tmp/redis/secret-key
+if [[ ! -f "tmp/redis/secret-key" ]]; then
+    mv tmp/secrets/redis-secret-key tmp/redis/secret-key
+fi
+
+# if we didn't need this temp secret, delete it
+if [[ -f "tmp/secrets/redis-secret-key" ]]; then
+    rm tmp/secrets/redis-secret-key
+fi
